Update move folders
