+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import MovieList from "./components/MovieList";
+import MovieList from "../components/MovieList";
 
 const App = () => (
   <main className="flex min-h-screen flex-col items-center p-24 bg-gradient-to-r from-slate-900 to-slate-700">
