+++ src/components/MovieItem.tsx
